add traits

+++ generic_types/src/main.rs
@@ -22,8 +22,8 @@ impl<T, U> Point<T, U> {
     }
 }
 
-impl Point<f64, f64> {
-    fn distance_from_origin(&self) -> f64 {
+impl Point<f32, f32> {
+    fn distance_from_origin(&self) -> f32 {
         (self.x.powi(2) + self.y.powi(2)).sqrt()
     }
 }
@@ -42,7 +42,7 @@ fn main() {
     let both_integer = Point { x: 5, y: 10 };
     println!("x: {}, y: {}", both_integer.x, both_integer.y);
     println!("x: {}, y: {}", both_integer.x(), both_integer.y());
-    let both_float = Point { x: 1.0, y: 4.0 };
+    let both_float = Point { x: 1.0 as f32, y: 4.0 as f32 };
     println!("Distance from origin: {}", both_float.distance_from_origin());
     let _integer_and_float = Point { x: 5, y: 4.0 };
